+++ LAKMobile/src/screens/index.tsx
@@ -1,3 +1,5 @@
 export * from './SignupScreen';
 export * from './LoginScreen';
 export * from './AddJob';
+export * from './ForgotPassword';
+export * from './OTP';
